+++ gamma_dashboard/static/dashboard/js/app/src/components/dashboard/rowBlock/progress-badge/ProgressBadge.jsx
@@ -1,25 +1,17 @@
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useEffect, useRef } from 'react';
 import PropTypes from 'prop-types';
-import { usePopper } from 'react-popper';
-import classNames from 'classnames';
+import { OverlayTrigger, Popover, useToggle } from '@openedx/paragon';
 
-import RowBlockItemPopup from '../RowBlockItemPopup';
-import ProgressBadgeDiagram from './ProgressBadgeDiagram';
-import ProgressBadgeFigure from './ProgressBadgeFigure';
+import { ProgressPropType } from '../../../propTypes';
+import { PopoverContent } from './popover-content';
 import { getTotalProgress, updateProgressView } from './utils';
-
-const DEFAULT_POPPER_STYLES = {
-  placement: 'auto',
-  scroll: false,
-};
+import ProgressBadgeItem from './ProgressBadgeItem';
 
 const ProgressBadge = ({ data, center, children }) => {
-  const [showPopup, setShowPopup] = useState(false);
-  const [referenceElement, setReferenceElement] = useState(null);
-  const [popperElement, setPopperElement] = useState(null);
-  const { styles, attributes } = usePopper(referenceElement, popperElement, DEFAULT_POPPER_STYLES);
-
-  const progressRef = useRef(null);
+  const [isPopoverOpen, openPopover, closePopover] = useToggle(false);
+  const popoverElementRef = useRef(null);
+  const badgeElementRef = useRef(null);
+  const progressElementRef = useRef(null);
 
   const {
     title,
@@ -27,21 +19,13 @@ const ProgressBadge = ({ data, center, children }) => {
     progress,
     points,
     statusPoints,
+    dependencies = [],
   } = data;
 
-  const dependencies = data.dependencies || [];
   const { hasPopup, totalProgressPercent } = getTotalProgress(data);
-
-  const itemProps = hasPopup ? {
-    onMouseEnter: () => setShowPopup(true),
-    onMouseLeave: () => setShowPopup(false),
-  } : {};
 
   const popupProps = hasPopup
     ? {
-      style: showPopup
-        ? { visibility: 'visible', opacity: 1 }
-        : {},
       data: {
         badgeDependencies: dependencies,
         statusDependency: data.statusDependency || null,
@@ -52,76 +36,106 @@ const ProgressBadge = ({ data, center, children }) => {
     } : null;
 
   useEffect(() => {
-    updateProgressView(progressRef.current, totalProgressPercent);
+    updateProgressView(progressElementRef.current, totalProgressPercent);
   }, [totalProgressPercent]);
 
-  return (
-    <li
-      className={classNames('progress-badge', {
-        'progress-badge-center': center,
-        'progress-badge-completed': !hasPopup,
-      })}
-      data-testid="progress-badge"
-      /* eslint-disable-next-line jsx-a11y/no-noninteractive-element-to-interactive-role */
-      role="button"
-      aria-haspopup="true"
-      aria-expanded={hasPopup}
-      tabIndex={0}
-      {...itemProps}
-    >
-      {hasPopup ? (
-        <>
-          <p className="total-progress-percent" data-testid="total-progress-percent">
-            {totalProgressPercent}%
-          </p>
-          <ProgressBadgeDiagram totalProgressPercent={totalProgressPercent} progressRef={progressRef}>
-            <ProgressBadgeFigure imageSrc={imageSrc} setReferenceElement={setReferenceElement} isDisabled />
-          </ProgressBadgeDiagram>
-        </>
-      ) : (
-        <ProgressBadgeFigure imageSrc={imageSrc} setReferenceElement={setReferenceElement} />
-      )}
-      <div className="progress-badge-title" data-testid="progress-badge-title">
-        {title}
-      </div>
-      {hasPopup && popupProps && (
-        <RowBlockItemPopup
-          ref={setPopperElement}
+  if (!hasPopup && !popupProps) {
+    return (
+      <ProgressBadgeItem
+        title={title}
+        imageSrc={imageSrc}
+        badgeRef={badgeElementRef}
+        center={center}
+        progressRef={progressElementRef}
+        totalProgressPercent={totalProgressPercent}
+      >
+        {children}
+      </ProgressBadgeItem>
+    );
+  }
+
+  const handleBadgeKeyDown = ({ key }) => {
+    if (key === 'Enter') {
+      openPopover();
+      requestAnimationFrame(() => {
+        const firstElement = popoverElementRef.current?.firstElementChild;
+        if (firstElement) {
+          firstElement.focus({ preventScroll: true });
+        }
+      });
+    }
+  };
+
+  const handlePopoverKeyDown = ({ key }) => {
+    if (key === 'Escape' || key === 'Tab') {
+      closePopover();
+      badgeElementRef.current?.focus();
+    }
+  };
+
+  const handlePopoverOpen = () => openPopover();
+  const handlePopoverClose = () => closePopover();
+
+  const renderPopover = (props) => (
+    <div ref={popoverElementRef}>
+      <Popover
+        id="popover-positioned-popover"
+        datatest-id="popover-positioned-popover"
+        tabIndex={0}
+        onMouseEnter={handlePopoverOpen}
+        onMouseLeave={handlePopoverClose}
+        onFocus={handlePopoverOpen}
+        onBlur={handlePopoverClose}
+        onKeyDown={handlePopoverKeyDown}
+        {...props}
+      >
+        <Popover.Title as="h3">
+          <strong>{title}</strong>
+        </Popover.Title>
+        <PopoverContent
           title={title}
-          style={{ ...styles.popper, ...popupProps.style }}
           data={popupProps.data}
-          {...attributes.popper}
         />
-      )}
-      {children}
-    </li>
+      </Popover>
+    </div>
+  );
+
+  return (
+    <OverlayTrigger
+      show={isPopoverOpen}
+      key="progress-badge-popover"
+      overlay={renderPopover}
+      placement="auto"
+    >
+      <ProgressBadgeItem
+        title={title}
+        imageSrc={imageSrc}
+        badgeRef={badgeElementRef}
+        center={center}
+        hasPopup
+        progressRef={progressElementRef}
+        totalProgressPercent={totalProgressPercent}
+        role="button"
+        aria-haspopup={hasPopup}
+        aria-expanded={isPopoverOpen}
+        tabIndex={0}
+        onMouseEnter={handlePopoverOpen}
+        onMouseLeave={handlePopoverClose}
+        onBlur={handlePopoverClose}
+        onKeyDown={handleBadgeKeyDown}
+      >
+        {children}
+      </ProgressBadgeItem>
+    </OverlayTrigger>
   );
 };
-
-const EventPropType = PropTypes.shape({
-  count: PropTypes.number.isRequired,
-  goal: PropTypes.number.isRequired,
-});
 
 ProgressBadge.propTypes = {
   slug: PropTypes.string,
   data: PropTypes.shape({
     title: PropTypes.string,
     url: PropTypes.string,
-    progress: PropTypes.shape({
-      edxForumCommentCreated: EventPropType,
-      openassessmentblockSaveSubmission: EventPropType,
-      edxBookmarkAdded: EventPropType,
-      edxForumThreadCreated: EventPropType,
-      edxForumResponseCreated: EventPropType,
-      edxForumThreadVoted: EventPropType,
-      stopVideo: EventPropType,
-      edxCertificateCreated: EventPropType,
-      edxGradesProblemSubmitted: EventPropType,
-      edxCourseEnrollmentActivated: EventPropType,
-      problemCheck: EventPropType,
-      problemGraded: EventPropType,
-    }),
+    progress: PropTypes.shape(ProgressPropType),
     done: PropTypes.bool,
     active: PropTypes.bool,
     points: PropTypes.number,