+++ gamma_dashboard/static/dashboard/js/app/src/components/dashboard/rowBlock/progress-badge/popover-content/ProgressBlock.jsx
@@ -1,0 +1,29 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+
+import { capitalizeFirstLetter } from '../../../../../utils';
+
+const ProgressBlock = ({ progressValues }) => {
+  const { count = 0, goal = 0, title: progressTitle = '' } = progressValues || {};
+
+  return (
+    <ul>
+      <li>
+        <span>
+          {`${Math.min(count, goal)}/${goal}`}
+        </span>
+        {capitalizeFirstLetter(progressTitle)}
+      </li>
+    </ul>
+  );
+};
+
+ProgressBlock.propTypes = {
+  progressValues: PropTypes.shape({
+    count: PropTypes.number,
+    goal: PropTypes.number,
+    title: PropTypes.string,
+  }).isRequired,
+};
+
+export default ProgressBlock;