+++ gamma_dashboard/static/dashboard/js/app/src/components/leaderboard/LeaderboardTable.jsx
@@ -6,6 +6,7 @@ import { Loader } from '../generic';
 
 import { getUserStatus } from '../../utils';
 import { useTranslate } from '../../i18n/utils';
+import { ProgressPropType, StatusPropType } from '../propTypes';
 
 const LeaderboardTable = ({
   rank, profiles, systemStatuses, delimiter,
@@ -36,22 +37,6 @@ const LeaderboardTable = ({
   );
 };
 
-const EventPropType = PropTypes.shape({
-  count: PropTypes.number.isRequired,
-  goal: PropTypes.number.isRequired,
-});
-
-const StatusPropType = PropTypes.shape({
-  statusUid: PropTypes.string.isRequired,
-  slug: PropTypes.string.isRequired,
-  title: PropTypes.string.isRequired,
-  description: PropTypes.string,
-  active: PropTypes.bool.isRequired,
-  statusPoints: PropTypes.number.isRequired,
-  color: PropTypes.string.isRequired,
-  url: PropTypes.string.isRequired,
-});
-
 LeaderboardTable.propTypes = {
   profiles: PropTypes.arrayOf(
     PropTypes.shape({
@@ -63,26 +48,13 @@ LeaderboardTable.propTypes = {
           title: PropTypes.string,
           description: PropTypes.string,
           done: PropTypes.bool,
-          progress: PropTypes.shape({
-            edxForumCommentCreated: EventPropType,
-            openassessmentblockSaveSubmission: EventPropType,
-            edxBookmarkAdded: EventPropType,
-            edxForumThreadCreated: EventPropType,
-            edxForumResponseCreated: EventPropType,
-            edxForumThreadVoted: EventPropType,
-            stopVideo: EventPropType,
-            edxCertificateCreated: EventPropType,
-            edxGradesProblemSubmitted: EventPropType,
-            edxCourseEnrollmentActivated: EventPropType,
-            problemCheck: EventPropType,
-            problemGraded: EventPropType,
-          }),
+          progress: PropTypes.shape(ProgressPropType),
           url: PropTypes.string,
         }),
       ),
     }),
   ).isRequired,
-  systemStatuses: PropTypes.arrayOf(StatusPropType).isRequired,
+  systemStatuses: PropTypes.arrayOf(PropTypes.shape(StatusPropType)).isRequired,
   rank: PropTypes.number.isRequired,
   delimiter: PropTypes.number,
 };